# openapi.yaml
...
openapi: 3.1.0
info:
  title: Asana Connector API
  version: 1.0.0
  summary: HTTP API over the DOO Asana connector.
  description: |-
    The HTTP adapter over the Asana connector core.

    All five actions are exposed through a single generic route,
    `POST /api/actions/{actionId}`, so there are no per-action handlers
    that can drift from the connector.

    ## Authentication

    Credentials live server-side only. The browser never holds an Asana
    token: it calls this API, and the server attaches authentication.
    Configure a Personal Access Token or an OAuth app via environment
    variables — see `.env.example`.

    ## Write safety

    The three write actions require `approved: true`. Failures on
    non-idempotent writes (create, comment) are reported with
    `retryStrategy: "manual_with_idempotency_key"` and are never retried
    automatically, because a timed-out create may already have succeeded.

    ## Errors

    Every failure uses one normalized envelope with a stable `code`,
    a `requestId`, and retry classification. Stack traces and tokens are
    never included.
  contact:
    name: Idrees Khaled
  license:
    name: MIT
servers:
  - url: http://localhost:8787
    description: Local development
  - url: "{baseUrl}"
    description: Deployed
    variables:
      baseUrl:
        default: http://localhost:8787
...
paths:
...
  /api/connector/schemas/{actionId}:
    get:
      operationId: getActionSchema
      summary: JSON Schema for one action
      tags:
        - connector
      parameters:
        - name: actionId
          in: path
          required: true
          schema:
            type: string
            enum:
              - asana.list_projects
              - asana.list_project_tasks
              - asana.create_task
              - asana.update_task
              - asana.add_comment
      responses:
        "200":
          description: Input and output JSON Schema.
...
components:
...
  securitySchemes:
    serverSideCredential:
      type: apiKey
      in: header
      name: x-not-used
      description: This API does not accept Asana credentials from the client. The server holds the
        Personal Access Token or OAuth tokens and attaches them upstream, so the browser never
        handles a secret.